fix: decode html entities in ai-assist API responses
