+++ src/agentTester.ts
@@ -12,6 +12,7 @@ import { ComponentSetBuilder, DeployResult, FileProperties, RequestStatus } from
 import { parse, stringify } from 'yaml';
 import { XMLBuilder, XMLParser } from 'fast-xml-parser';
 import { MaybeMock } from './maybe-mock';
+import { decodeHtmlEntities } from './utils';
 
 export type TestStatus = 'NEW' | 'IN_PROGRESS' | 'COMPLETED' | 'ERROR' | 'TERMINATED';
 
@@ -389,49 +390,6 @@ export function normalizeResults(results: AgentTestResultsResponse): AgentTestRe
       })),
     })),
   };
-}
-
-/**
- * Clean a string by replacing HTML entities with their respective characters.
- *
- * @param str - The string to clean.
- * @returns The cleaned string with all HTML entities replaced with their respective characters.
- */
-function decodeHtmlEntities(str: string = ''): string {
-  const entities: { [key: string]: string } = {
-    '&quot;': '"',
-    '&apos;': "'",
-    '&amp;': '&',
-    '&lt;': '<',
-    '&gt;': '>',
-    '&#39;': "'",
-    '&deg;': '°',
-    '&nbsp;': ' ',
-    '&ndash;': '–',
-    '&mdash;': '—',
-    '&rsquo;': '’',
-    '&lsquo;': '‘',
-    '&ldquo;': '“',
-    '&rdquo;': '”',
-    '&hellip;': '…',
-    '&trade;': '™',
-    '&copy;': '©',
-    '&reg;': '®',
-    '&euro;': '€',
-    '&pound;': '£',
-    '&yen;': '¥',
-    '&cent;': '¢',
-    '&times;': '×',
-    '&divide;': '÷',
-    '&plusmn;': '±',
-    '&micro;': 'µ',
-    '&para;': '¶',
-    '&sect;': '§',
-    '&bull;': '•',
-    '&middot;': '·',
-  };
-
-  return str.replace(/&[a-zA-Z0-9#]+;/g, (entity) => entities[entity] || entity);
 }
 
 async function jsonFormat(results: AgentTestResultsResponse): Promise<string> {
