+++ src/utils.ts
@@ -1,0 +1,49 @@
+/*
+ * Copyright (c) 2025, salesforce.com, inc.
+ * All rights reserved.
+ * Licensed under the BSD 3-Clause license.
+ * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
+ */
+
+/**
+ * Clean a string by replacing HTML entities with their respective characters.
+ *
+ * @param str - The string to clean.
+ * @returns The cleaned string with all HTML entities replaced with their respective characters.
+ */
+export const decodeHtmlEntities = (str: string = ''): string => {
+  const entities: { [key: string]: string } = {
+    '&quot;': '"',
+    '&apos;': "'",
+    '&amp;': '&',
+    '&lt;': '<',
+    '&gt;': '>',
+    '&#39;': "'",
+    '&deg;': '°',
+    '&nbsp;': ' ',
+    '&ndash;': '–',
+    '&mdash;': '—',
+    '&rsquo;': '’',
+    '&lsquo;': '‘',
+    '&ldquo;': '“',
+    '&rdquo;': '”',
+    '&hellip;': '…',
+    '&trade;': '™',
+    '&copy;': '©',
+    '&reg;': '®',
+    '&euro;': '€',
+    '&pound;': '£',
+    '&yen;': '¥',
+    '&cent;': '¢',
+    '&times;': '×',
+    '&divide;': '÷',
+    '&plusmn;': '±',
+    '&micro;': 'µ',
+    '&para;': '¶',
+    '&sect;': '§',
+    '&bull;': '•',
+    '&middot;': '·',
+  };
+
+  return str.replace(/&[a-zA-Z0-9#]+;/g, (entity) => entities[entity] || entity);
+};
